refactor: group skill pages under pages/skill/ (#104)
Pure file-move + import update; no route or behavior change.

- 9 skill-related pages moved into ornn-web/src/pages/skill/:
  CreateSkillFreePage, CreateSkillFromGitHubPage, CreateSkillGenerativePage,
  CreateSkillGuidedPage, EditSkillPage, MySkillsPage, SkillAuditHistoryPage,
  SkillDetailPage, UploadSkillPage.
- New pages/skill/index.ts barrel re-exports them.
- App.tsx lazy imports updated to the new paths.
- pages/admin/ and the flat top-level pages (LandingPage / LoginPage /
  PlaygroundPage / NotificationsPage / MySharesPage / ReviewsPage etc.)
  remain where they are — only skill-domain pages were grouped this round.
