+++ ornn-web/src/pages/skill/CreateSkillGenerativePage.tsx
