+++ ornn-web/src/pages/skill/SkillDetailPage.tsx
