+++ ornn-web/src/pages/skill/index.ts
@@ -1,0 +1,16 @@
+/**
+ * Skill pages barrel — keeps `App.tsx` imports tidy and gives consumers a
+ * single place to discover the routes that live under `/skills/...`.
+ *
+ * @module pages/skill
+ */
+
+export { CreateSkillFreePage } from "./CreateSkillFreePage";
+export { CreateSkillFromGitHubPage } from "./CreateSkillFromGitHubPage";
+export { CreateSkillGenerativePage } from "./CreateSkillGenerativePage";
+export { CreateSkillGuidedPage } from "./CreateSkillGuidedPage";
+export { EditSkillPage } from "./EditSkillPage";
+export { MySkillsPage } from "./MySkillsPage";
+export { SkillAuditHistoryPage } from "./SkillAuditHistoryPage";
+export { SkillDetailPage } from "./SkillDetailPage";
+export { UploadSkillPage } from "./UploadSkillPage";
